DO NOT MERGE Fix issue with bubble stack & bubble data expanded bit being out of sync
If onBackPressed was used while the stack was expanded, it was calling
setExpanded on BubbleStackView (BSV) and NOT BubbleData. This results in
BSV and BubbleData being out of sync which can lead to incorrect behavior.
(e.g. using individual bubble dismiss method instead of the stack dismiss
method).

Test: atest BubblesTest NewNotifPipelineTest
Fixes: 217361258
Change-Id: I748fdd49626acaa0e986a9d8038039cada6faa60
Merged-In: I748fdd49626acaa0e986a9d8038039cada6faa60