Support restoring the original task ordering of transient-launches
Like legacy recents, we need to be able to restore the home task
back to where it was in the task-ordering when recents doesn't
actually switch to home.

This CL records the "restore-below" task during activity-start
and stores in the associated transition. There is a new WCT op
that will "restore" a container back to where it was.

Bug: 210014802
Test: launch a task chain (task followed by task), then go to
      recents, then restore the top task. observe task order
      remains as it was before recents.
Change-Id: I9f9da7d35ffe69099e6431b7a0d5a527e33c5a74