Update READ_CLIPBOARD_IN_BACKGROUND to be grantable via roles
ClipboardService currently has a special case to allow Android System
Intelligence to read the clipboard in the background. The special case
is implemented by checking if the calling package is the provider of
ContentCaptureService. Rather than having to maintain this special case,
we want to reuse the existing READ_CLIPBOARD_IN_BACKGROUND permission
by allowing it to be associated with roles. This way the permission will
be associated with the appropriate role held by Android System
Intelligence and the special case can be removed from ClipboardService.
This also decouples which PCC package provides ContentCaptureService vs
which PCC package should be allowed to access the clipboard.

The change to remove the special case logic from ClipboardService will
be done as a separate CL.

Bug: 216314180
Test: atest PermissionPolicyTest
Change-Id: I68abf8c200f6827fc350ab2f990eb78d9d850c8b