Optimize the handling of disabling doze state.
When we are handling the callback for doze-disabled
state, we don't have to recompute the network rules
for all uids anymore since we started keeping track of
individual restriction state for each uid. We can
remove the doze restriction state from only blocked uids.

We can also do the same when we are handling the
doze-enabled state (and even other restrictions) but since
NPMS doesn't get callbacks for every uid procstate change,
it is possible we can miss blocking certain uids network.
So, for now limiting the change to only doze-disabled state.

Bug: 206518114
Test: tests/cts/hostside/src/com/android/cts/net/HostsideRestrictBackgroundNetworkTests.java
Change-Id: I53e40af53aa00ff493e8e5050908220ae4dad729